data/imageCaptioningCOCO/requirements.txt
torch
nltk
pycocotools
torchvision
pillow
pandas
pyarrow
fastparquet
